requirements.txt

flask
line-bot-sdk
google-api-python-client
google-auth
google-auth-oauthlib
apscheduler
